Move `WithClock` to `unit` build tagged file
