+++ events/queue/processor_unit.go
@@ -1,0 +1,27 @@
+//go:build unit
+// +build unit
+
+/*
+Copyright 2023 The Dapr Authors
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+    http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package queue
+
+import (
+	"k8s.io/utils/clock"
+)
+
+// WithClock sets the clock used by the processor. Used for testing.
+func (p *Processor[T]) WithClock(clock clock.Clock) *Processor[T] {
+	p.clock = clock
+	return p
+}
